Use some unused Status responses.
GitOrigin-RevId: b49d15a22918169b5e1a212e5858dc2e9bd2033e

+++ td/telegram/files/PartsManager.cpp
@@ -215,8 +215,8 @@ string PartsManager::get_bitmask() {
 }
 
 bool PartsManager::is_part_in_streaming_limit(int part_i) const {
-  auto offset_begin = static_cast<int64>(part_i * get_part_size());
-  auto offset_end = static_cast<int64>(offset_begin + get_part(part_i).size);
+  auto offset_begin = static_cast<int64>(part_i) * static_cast<int64>(get_part_size());
+  auto offset_end = offset_begin + static_cast<int64>(get_part(part_i).size);
 
   if (offset_begin >= get_expected_size()) {
     return false;
